Move sha256d tests to `test_hashutil.py`
