+++ tests/test_lafs.py
@@ -1,28 +1,12 @@
 from pathlib import Path
 
-import pytest
 import yaml
 
 from allmydata.util.base32 import b2a
-from allmydata.util.hashutil import (
-    _SHA256d_Hasher,  # XXX
-    ssk_pubkey_fingerprint_hash,
-    ssk_writekey_hash,
-)
+from allmydata.util.hashutil import ssk_pubkey_fingerprint_hash, ssk_writekey_hash
 from cryptography.hazmat.primitives import serialization
 
 from lafs import derive_mutable_uri
-
-
-@pytest.mark.parametrize("data, expected", [
-    (b"", b"lx3obytwcnm5gcucoucy4km7zqbycu2fix2vz5b6igmd6xkmsrla"),
-    (b"test", b"svgvusp5odm3rpg3gxjfejtyfgkx67xx7jwhj6eedg64l2bcbh2a"),
-])
-def test_sha256d(data, expected) -> None:
-    hasher = _SHA256d_Hasher()
-    hasher.update(data)
-    digest = hasher.digest()
-    assert b2a(digest) == expected
 
 
 def py_derive_mutable_uri(private_key_pem: str, format: str) -> str:
